refactor: change some code, moved functions to utils
